+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [2025.2.4.2]
 
 ### `xDebugger` enhancements
+- Re-implemented model debugger to lazy-evaluation of the fields [#1579](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1579)
 - Renderer is not always created first time [#1578](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1578)
 
 ## [2025.2.4.1]